Improve timeout behavoir of Countdown Timer and Stopwatch
Return to time display if in Timer or Sotpwatch mode for more than 60 minutes without any activity.
This does not apply if Timer or Stopwatch are running. In this case, the 60 minutes are counted from the moment when either feature is stopped.
Refactored code by simplifying conditional statements inside the menu state machine.